Cosmetic changes. Split off Helper and TestKML
Working towards having the elements in KML in alphabetical order.
