+++ src/main/scala/com/phasmidsoftware/kmldoc/Helper.scala
@@ -1,0 +1,146 @@
+package com.phasmidsoftware.kmldoc
+
+import com.phasmidsoftware.core.{Text, TryUsing}
+import com.phasmidsoftware.kmldoc.KmlEdit.editFeatures
+import com.phasmidsoftware.render._
+import com.phasmidsoftware.xml.Extractor
+import scala.reflect.ClassTag
+import scala.xml.NamespaceBinding
+
+object Helper
+
+/**
+ * Trait to define the property of owning features.
+ */
+trait HasFeatures {
+  val features: Seq[Feature]
+}
+
+/**
+ * Companion object HasFeatures provides utilities for editing objects that have a property of owning features.
+ */
+object HasFeatures {
+
+  /**
+   * Method to edit an object that has features.
+   *
+   * @param edit the edit to be applied.
+   * @param t    the object to edit.
+   * @param g    a function which takes a Seq[Feature] and creates an optional new copy of the T based on the provided Feature sequence.
+   * @tparam T the type of <code>t</code>.
+   * @return an optional copy of <code>t</code> with a (potentially) new set of features.
+   */
+  def editHasFeaturesToOption[T](t: T)(edit: KmlEdit)(g: Seq[Feature] => T): Option[T] = t match {
+    case h: HasFeatures => Some(g(editFeatures(edit, h.features)))
+    case _ => throw new Exception(s"editHasFeaturesToOption: parameter t does not extend HasFeatures: ${t.getClass}")
+  }
+
+}
+
+/**
+ * Trait representing an entity that has a name attribute.
+ *
+ * A `HasName` entity provides access to its name, defined as a `Text` type.
+ */
+trait HasName {
+  /**
+   * Retrieves the name as a `Text` value.
+   *
+   * @return the name represented as a `Text` instance.
+   */
+  def name: Text
+}
+
+/**
+ * Represents a binding for a KML object with its corresponding XML namespace binding.
+ *
+ * NOTE this is a bit of a mystery.
+ *
+ * A KML_Binding associates a KML instance with its XML NamespaceBinding to manage XML namespaces.
+ *
+ * @constructor Constructs a new KML_Binding object with a given KML instance and its NamespaceBinding.
+ * @param kml     the KML object representing the main geospatial data structure.
+ * @param binding the NamespaceBinding for managing XML namespaces linked to the KML.
+ */
+case class KML_Binding(kml: KML, binding: NamespaceBinding)
+
+/**
+ * Companion object for the `KML_Binding` case class.
+ * Provides implicit implementations for extracting and rendering `KML_Binding` instances.
+ *
+ * The object defines two implicit values:
+ * 1. An `Extractor[KML_Binding]` to parse an XML `Node` and construct a `KML_Binding` instance.
+ * 2. A `Renderer[KML_Binding]` to facilitate rendering of a `KML_Binding` instance to XML format.
+ *
+ * These implicits make the `KML_Binding` interoperable with XML nodes and compatible with processes
+ * that work with Extraction and Rendering mechanisms.
+ */
+object KML_Binding {
+  implicit val extractor: Extractor[KML_Binding] = Extractor {
+    node =>
+      Extractor.extract[KML](node) map (KML_Binding(_, node.scope))
+  }
+  implicit val renderer: Renderer[KML_Binding] = Renderer {
+    (t: KML_Binding, format: Format, stateR: StateR) =>
+      TryUsing(stateR.addAttribute(s"""${t.binding}"""))(rs => Renderer.render(t.kml, format, rs))
+  } ^^ "rendererKml_Binding"
+}
+
+/**
+ * Object KmlRenderers extends the Renderers base trait and serves as a collection
+ * of formatting utilities specifically for KML (Keyhole Markup Language) rendering.
+ *
+ * This object contains nested case classes and methods for handling formatting operations.
+ */
+object KmlRenderers extends Renderers {
+
+  /**
+   * FormatCoordinate is a case class for handling coordinate-specific formatting in KML rendering.
+   * It extends the BaseFormat abstraction, inheriting core formatting behavior while specializing in
+   * coordinate-related formatting operations.
+   *
+   * @param indents The number of indentation levels used for formatting.
+   */
+  case class FormatCoordinate(indents: Int) extends BaseFormat(indents) {
+    val name: String = "formatCoordinate"
+
+    /**
+     * Returns the current instance of the Format, as it represents the same formatting configuration.
+     *
+     * @return the current Format instance with unchanged characteristics.
+     */
+    def indent: Format = this
+
+    /**
+     * Formats the name based on the given state and an optional flag.
+     *
+     * @param open   an optional Boolean value. If Some(true), specific formatting is performed; otherwise, a default behavior is applied.
+     * @param stateR the rendering state represented by the StateR instance, which provides context for formatting.
+     * @tparam T a type parameter with an implicit ClassTag, allowing runtime inspection of type T.
+     * @return a formatted string based on the input parameters and formatting logic.
+     */
+    def formatName[T: ClassTag](open: Option[Boolean], stateR: StateR): String = open match {
+      case Some(true) =>
+        ""
+      case _ =>
+        ""
+    }
+
+    /**
+     * Determines the sequence string based on an optional Boolean input.
+     *
+     * @param open An Option of Boolean indicating the desired state. If Some(true), a newline with the specified indentation is returned.
+     *             If Some(false), an empty string is returned. If None, a single newline character is returned.
+     * @return A string representing the appropriate sequence based on the input parameter.
+     */
+    def sequencer(open: Option[Boolean]): String = open match {
+      case Some(true) => newline
+      case Some(false) => ""
+      case _ => "\n"
+    }
+  }
+
+  // TODO refactor the sequenceRendererFormatted method so that its parameter is a Format=>Format function.
+}
+
+case class KmlException(str: String) extends Exception(str)
